#1 #2 Centralized the logic of the SaveToDatabase job to accept a Data Manager object instead of a date so it applies the same code path when saving data to the database regardless of the data source.
Completed the process that saves leaderboard/entry data into the database for both XML and CSV processes.

Centralized the logic for the process that uploads leaderboard data to S3 into one unified job and modified all applicable console commands to dispatch this instead of the old ones.

Fixed various bugs encountered by code paths that are invoked for when XML is being saved.