#4 Created the Multi and Pipeline redis transaction components.
Refactored CallbackHandler for working with PHP 7.2 and general code cleanup.

Introduced the RecordQueue to serve as a base for processing batches of records.

Refactored InsertQueue so that it utilizes RecordQueue and CallbackHandler to reduce duplicated code.

Updated all methods that return an instance of InsertQueue to work with the refactored object.

Changed the Laravel alias for redis to be RedisClient to work with the Redis object provided by phpredis.