#5 Migrated the power ranking generation process.
Migrated the process that creates a power ranking entries partition table.

Modified the power ranking entries partition table structure so that all character specific data is contained in a jsonb field to dramatically reduce the number of columns needed to store the same data. Incorporated this change into the power ranking generation process and partition table creation process.

Dropped the default value of power_ranking_id and changed the incrementing of its value to be managed in code via calls to its sequence.

Migrated the CacheNames library for power rankings and refactored it for simplicity.

Added the GetById trait and added it to all applicable models.

Migrated the algorithm to generate points from a rank to the RankPoints model.

Fixed various bugs that came up during this process.