#6 Migrated the daily ranking generation process and its corresponding components.
 #14 #15 Migrated the processes to create power ranking and daily ranking partition tables.

Centralized the functionality found in ranking generation that auto selects a redis database that's not in use for processing into the DataSelector component.

Created the HasPartitions and HasTempTable traits to centralize a lot of duplicated code and updated all affected models to use these traits to reduce code duplication.

Removed the RedisTransation component that was accidentally added.