#16 Migrated the processes to import replays from Steam and save them to the database.
Added the SteamReplays object to the SteamDataManager library to handle interactions with replay files.

Created the seeds table and modified the steam_replays table to replace its seed field with a foreign key seed_id field.

Modified the steam replay import process to add all seeds to the seeds table.

Added functionality to generate a random seed to the Seeds model.

Removed auto incrementing for the primary keys of run_results and steam_replay_versions and added the HasManualSequence traits to their respective models.