#90 Removed 'steam' from leaderboard api endpoint urls.
 #35 Migrated the /leaderboards/xml api endpoint.

 #46 Migrated the /rankings/power/entries api endpoint.

 #47 Migrated the /rankings/power/score/entries api endpoint.

 #48 Migrated the /rankings/power/speed/entries api endpoint.

 #49 Migrated the /rankings/power/deathless/entries api endpoint.

 #50 Migrated the /rankings/daily/entries api endpoint.

 #91 Modified the power rankings generation process to include score,
win counts, and times for each character and leaderboard type.

Fixed various issues with the power ranking generation and caching processes.

Modified EntriesDataset to allow retrieval of and decoding of data streams for specified bytea fields prior to paginator caching.

Added functionality to EntriesDataset that sorts the retrieved data by a
specified callback.