#93 Refactored EntriesDataset into Dataset, a modular library with the index pagination and data retrieval separated out into their own modules.
Converted all api endpoints that used EntriesDataset to use Dataset instead.

Moved data sorting from the query orderBy method to Dataset sorting callbacks for leaderboard entries endpoints.

Modified SteamUsers::getApiReadQuery() so that it doesn't allow an optional array of steamids to be passed in to grab only those queries. This feature was unused.

Fixed a route issue with the /api/1/leaderboards/entries endpoint that caused the wrong controller and method to be invoked.

Removed EntriesDataset.