#94 Added steam_user_id to leaderboard_entries_* tables for significantly faster queries against it for the API.
Modifed the process to save leaderboard entries to the database and the process to cache leaderboard entries to use steam_user_id instead of steam_user_pb_id for index values.

Added fulltext searching to the steam_users table by adding a personaname_search_index field of type tsvector with a GIN index.

Implemented SteamUsers::getIdsBySearchTerm() that returns all steam_user_ids that match the given search term and caches the result for 5 minutes.

Modified the Sql index to utilize SteamUsers::getIdsBySearchTerm() for searching.

Updated caching in several spot to store data for 5 minutes instead of 1.