#399 Added the process to import leaderboard data from the legacy site's database.
Modified PostgresCursor to allow a different DB connection to be set.

Added the getLegacyImportQuery() method to the Leaderboards, LeaderboardSnapshots, and LeaderboardEntryDetails models.

Added an entry in the database config for the legacy site.

Updated the .env.example with constants for connecting to the legacy site.

Added the syncManualSequence method to the HasManualSequence trait to reset that table's sequence.